recipe_uploader/requirements.txt
google-generativeai
python-docx 